@@ -1,7 +1,6 @@
 package httpcache
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"reflect"
@@ -101,27 +100,8 @@ func (h *Handler) fetchUpstream(req *http.Request, next caddyhttp.Handler, key s
 
 	// Do the upstream fetching in background
 	go func(req *http.Request, response *Response) {
-		// Create a new context to avoid terminating the Next.ServeHTTP when the original
-		// request is closed. Otherwise if the original request is cancelled the other requests
-		// will see a bad response that has the same contents the first request has
-		updatedContext := context.Background()
 
-		// TODO: find a way to copy the origin request...
-		// The problem of cloning the context is that the original one has some values used by
-		// other middlewares. If those values are not present they break, #22 is an example.
-		// However there isn't a way to know which values a context has. I took the ones that
-		// I found on caddy code. If in a future there are new ones this might break.
-		// In that case this will have to change to another way
-		for _, key := range contextKeysToPreserve {
-			value := req.Context().Value(key)
-			if value != nil {
-				updatedContext = context.WithValue(updatedContext, key, value)
-			}
-		}
-
-		updatedReq := req.WithContext(updatedContext)
-
-		upstreamError := next.ServeHTTP(response, updatedReq)
+		upstreamError := next.ServeHTTP(response, req)
 		errChan <- upstreamError
 		response.Close()
 
